feat: Migrate VSCode settings plugin to mixin architecture
Complete Phase 2 pilot migration demonstrating the new architecture:

**Code Reduction Achieved:**
- 53% reduction: 358 → 168 lines of code
- Eliminated all boilerplate configuration handling
- Removed duplicate validation logic
- Simplified plugin structure drastically

**New Implementation:**
- VSCodeCore: 66 lines (pure business logic)
- Registration: 1 line (automatic mixin integration)
- Tests: 95 lines (comprehensive coverage maintained)

**Architecture Benefits:**
- Type-safe configuration through StandardConfig
- Automatic validation with helpful error messages
- Built-in restore functionality via FilesMixin
- Consistent interface patterns across all settings plugins
- Zero duplication of common plugin patterns

**Features Maintained:**
- Platform-specific settings directories (macOS/Windows/Linux)
- JSON/JSONC file validation
- Comprehensive test coverage
- Plugin auto-registration system
- Full restore functionality with dry-run support

Next: Ready for Phase 3 full migration of remaining plugins.
